Código arreglado (todos tienen que probarlo)

+++ include/csv_reader.h
@@ -9,10 +9,10 @@
 #define MAX_ENTRIES 100
 
 typedef struct {
-    int pizza_id;
-    int order_id;
-    char pizza_name_id[50];	
-    int quantity;
+    float pizza_id;
+    float order_id;
+    char pizza_name_id[50];    
+    float quantity;
     char order_date[11];
     char order_time[9];
     float unit_price;
@@ -21,7 +21,6 @@ typedef struct {
     char pizza_category[20];
     char pizza_ingredients[100];
     char pizza_name[50];
-
 } Venta;
 
 int leer_csv(const char *filename, Venta ventas[], int *total_ventas);
